refactor: reorganize component imports and remove unused sidebar components
