+++ src/components/react/MainControlsBar/AreaDefinition.tsx
